agentic-security/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentic-security"
version = "0.1.0"
description = "Agentic Identity Framework abstracting Keycloak/OIDC complexity for AI Agents"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
keywords = ["security", "oidc", "keycloak", "agents", "ai"]
authors = [
  {name = "Antigravity", email = "antigravity@example.com"},
]
dependencies = [
  "fastapi",
  "python-keycloak",
  "httpx",
  "pydantic",
]

[project.optional-dependencies]
test = [
  "pytest",
  "pytest-mock",
  "pytest-cov",
  "python-keycloak", 
  "httpx"
]

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_security"]
